requirements.txt
tensorflow
flask
pillow
werkzeug
numpy
